Move habitable zone renderable to space module
