+++ modules/exoplanets/CMakeLists.txt
@@ -27,7 +27,6 @@ include(${OPENSPACE_CMAKE_EXT_DIR}/module_definition.cmake)
 set(HEADER_FILES
     exoplanetshelper.h
     exoplanetsmodule.h
-    rendering/renderablehabitablezone.h
     rendering/renderableorbitdisc.h
     tasks/exoplanetsdatapreparationtask.h
 )
@@ -37,7 +36,6 @@ set(SOURCE_FILES
     exoplanetshelper.cpp
     exoplanetsmodule.cpp
     exoplanetsmodule_lua.inl
-    rendering/renderablehabitablezone.cpp
     rendering/renderableorbitdisc.cpp
     tasks/exoplanetsdatapreparationtask.cpp
 )
