+++ modules/space/include.cmake
@@ -1,1 +1,4 @@
 set(DEFAULT_MODULE ON)
+set (OPENSPACE_DEPENDENCIES
+  base
+)
